Rip out buffer and widgets
I've decided to use the "taffy" crate for layouting. To make the API
somewhat bearable to use, I've decided to perform all rendering in a
good ol' rgb image buffer (in this case, an ImageBuffer from the "image"
crate). The pixels will be converted to their final representation just
before being sent to whatever device will be displaying them.

I've removed the buffer and the color traits because they're no longer
necessary. I've also removed the existing widgets because the widget
system will have to be redesigned to work well with taffy.